Rename httpServer package to httpserver. Add server package
