+++ pom.xml
@@ -40,7 +40,7 @@
                         <manifest>
                             <addClasspath>true</addClasspath>
                             <classpathPrefix>lib/</classpathPrefix>
-                            <mainClass>httpServer.App</mainClass>
+                            <mainClass>httpserver.App</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
