--- src/main/java/httpServer/ReadResponseException.java
@@ -1,10 +1,0 @@
-package httpServer;
-
-import java.io.IOException;
-
-class ReadResponseException extends RuntimeException {
-
-    ReadResponseException(Exception cause) {
-        super(cause);
-    }
-}
