+++ src/main/java/httpserver/App.java
@@ -1,4 +1,7 @@
-package httpServer;
+package httpserver;
+
+import httpserver.server.CurrentThreadExecutor;
+import httpserver.server.Server;
 
 import java.net.ServerSocket;
