+++ src/main/java/httpserver/ReadResponseException.java
@@ -1,0 +1,8 @@
+package httpserver;
+
+public class ReadResponseException extends RuntimeException {
+
+    ReadResponseException(Exception cause) {
+        super(cause);
+    }
+}
