+++ src/main/java/httpserver/ResponseHandler.java
@@ -1,4 +1,4 @@
-package httpServer;
+package httpserver;
 
 import java.io.PrintWriter;
