+++ src/main/java/httpserver/server/ClientHandler.java
@@ -1,4 +1,6 @@
-package httpServer;
+package httpserver.server;
+
+import httpserver.RequestResponseHandler;
 
 import java.io.BufferedReader;
 import java.io.IOException;
