+++ src/main/java/httpserver/server/ClientInputOutputException.java
@@ -1,4 +1,4 @@
-package httpServer;
+package httpserver.server;
 
 public class ClientInputOutputException extends RuntimeException {
     public ClientInputOutputException(Exception cause) {
