+++ src/main/java/httpserver/server/ClientSocketException.java
@@ -1,4 +1,4 @@
-package httpServer;
+package httpserver.server;
 
 public class ClientSocketException extends RuntimeException {
     ClientSocketException(Exception cause) {
