+++ src/test/java/httpserver/RequestHandlerTest.java
@@ -1,4 +1,4 @@
-package httpServer;
+package httpserver;
 
 import org.junit.jupiter.api.Test;
