+++ src/test/java/httpserver/RequestResponseHandlerTest.java
@@ -1,4 +1,4 @@
-package httpServer;
+package httpserver;
 
 import org.junit.jupiter.api.Test;
