+++ src/test/java/httpserver/RouteHandlerTest.java
@@ -1,4 +1,4 @@
-package httpServer;
+package httpserver;
 
 import org.junit.jupiter.api.Test;
