+++ src/test/java/httpserver/server/ConsoleWriterTest.java
@@ -1,4 +1,4 @@
-package httpServer;
+package httpserver.server;
 
 import org.junit.jupiter.api.Test;
