+++ src/test/java/httpserver/server/ServerTest.java
@@ -1,4 +1,4 @@
-package httpServer;
+package httpserver.server;
 
 import org.junit.jupiter.api.Test;
